+++ project_cylon/TestServer.py
@@ -12,7 +12,7 @@ class TestServer(object):
     @cherrypy.tools.allow(methods=['GET'])
     def index(self):
         filepath, filename = os.path.split(__file__)
-        filename = os.path.join(filepath, "public", "test-runner.html")
+        filename = os.path.join(filepath, "public", "test-runner-new.html")
 
         htmlfile = open(filename, 'r')
         return htmlfile.read()