+++ project_cylon/api/ApiReport.py
@@ -31,6 +31,50 @@ class ApiReport(object):
             filepath, filename = os.path.split(file)
             filename = filename.replace('.xml', '.html')
 
-            info['files'].append({ 'filename': filename, 'status': testsuite.status })
+            item = {
+                'filename': filename,
+                'status': testsuite.status,
+                'passed': testsuite.passed,
+                'failed': testsuite.failed,
+                'total': testsuite.tests,
+                'duration': testsuite.time
+            }
 
+            info['files'].append(item)
+
+            #info['files'].append({ 'filename': filename, 'status': testsuite.status })
+
+        return json.dumps(info)
+
+    @cherrypy.expose
+    @cherrypy.tools.allow(methods=['GET'])
+    def mock_info(self):
+        info = {
+            'path': '/project/reports/html',
+            'files': [
+                {
+                    'filename:': 'report-passed.html',
+                    'status': 'passed',
+                    'passed': 12,
+                    'failed': 3,
+                    'total': 15,
+                    'duration': 10.1234
+                },
+                {
+                    'filename': 'report-failed.html',
+                    'status': 'failed',
+                    'passed': 12,
+                    'failed': 3,
+                    'total': 15,
+                    'duration': 10.1234
+                },
+                {
+                    'filename': 'report-skipped.html',
+                    'passed': 12,
+                    'failed': 3,
+                    'total': 15,
+                    'duration': 10.1234
+                }
+            ]
+        }
         return json.dumps(info)