+++ project_cylon/public/js/run-panel.js
@@ -1,0 +1,11 @@
+
+function RunPanel()
+{
+    this.render = function() {
+        return this.renderPanel();
+    }
+
+    this.renderPanel = function() {
+        
+    }
+}